requirements.txt
cycler==0.11.0
fonttools==4.38.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.21.6
packaging==22.0
Pillow==9.3.0
potrace @ file:///root/ShareEmoji/pypotrace
pyparsing==3.0.9
python-dateutil==2.8.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
svgwrite==1.4.3
threadpoolctl==3.1.0
typing-extensions==4.4.0
